alteraçoes em repository estoque

+++ src/main/java/cetam/projeto02grupo04/repository/EstoqueRepository.java
@@ -3,5 +3,5 @@ package cetam.projeto02grupo04.repository;
 import cetam.projeto02grupo04.model.Estoque;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstoqueRepository extends JpaRepository<Estoque, Long> {
+public interface EstoqueRepository extends JpaRepository<Estoque, Integer> {
 }
